feat(cmd): code cleaned
